handle CopyRect updates for multicast clients via unicast.
This way rfbSendMulticastFramebufferUpdate() can be a lot more
simple, we dont't have to modify the CopyRect functions and it
really doesn't make sense sending CopyRect via multicast anyway.

Semantics right now are that we assume that the server knows
what it's doing, so a CopyRect destination region is removed
from the region sent as ordinary pixel data. For this to work,
all multicast clients must support CopyRect. If they don't, it
is not used.