httpd: fix sending of binary data such as images.
We do this simply by omitting the content-type and let the browser
decide upon the mime-type of the sent file. Only exception is
'index.vnc', where we do set the content-type since some browsers
fail to detect it's html when it's ending in '.vnc'

Also, remove superfluous #defines. We close the connection always.