Fix build error when libpng is available, but libjpeg is not.
The png stuff in tight.c depends on code in tight.c that uses libjpeg
features. We could probably seperate that, but for now the dependency
for 'tight' goes:

PNG depends on JPEG depends on ZLIB.

This is reflected in Makefile.am now.

NB: Building tight.c with JPEG but without PNG is still possible,
    but nor the other way around.