Fix an issue that affects the existing Tight encoder as well as the newly-implemented Turbo encoder.
The issue is that, when using the current libvncserver source, it is impossible to disable Tight JPEG encoding.
The way Tight/Turbo viewers disable JPEG encoding is by simply not sending the Tight quality value, causing the
server to use the default value of -1.  Thus, cl->tightQualityLevel has to be set to -1 prior to processing the
encodings message for this mechanism to work.  Similarly, it is not guaranteed that the compress level will be
set in the encodings message, so it is set to a default value prior to processing the message.