Ship the required macros in the m4/ directory.
This is recommended practice as per
https://www.gnu.org/software/automake/manual/html_node/Local-Macros.html.

It fixes the problem that arose when one of the maintainers could not build LibVNCServer
after https://github.com/LibVNC/libvncserver/pull/38 was merged.

Symptoms included

  checking whether make sets $(MAKE)... yes
  ./configure: line 2481: syntax error near unexpected token `rfb/rfbconfig.h'
  ./configure: line 2481: `AX_PREFIX_CONFIG_H(rfb/rfbconfig.h)'

until autoconf-archive was installed (which was a previously unmentioned
requirement for Pull Request #38) – this is not always an option, in particular
when the project needs to be built using a system-wide autoconf installation
that cannot be modified easily by the developer.