Limit malloc() size to SIZE_MAX - 1 bytes
This ammends a 502821828ed00b4a2c4bef90683d0fd88ce495de fix for a heap
out-of-bound write access in rfbProcessFileTransferReadBuffer() when
reading a transfered file content in a server. The former fix did not
work on platforms with a 32-bit size_t type (expected by malloc()) or
with a 32-bit int type (expected by rfbReadExact());

It also fixes the same issue in InitialiseRFBConnection() when
reading a display name in a client.

It also corrects a cast in HandleRFBServerMessage() when a client
reads a cut text from a server. This allocation is already guarded
since commit c5ba3fee85a7ecbbca1df5ffd46d32b92757bc2a.

CVE-2018-15127
<https://github.com/LibVNC/libvncserver/issues/243>
<https://github.com/LibVNC/libvncserver/issues/273>