LibVNCServer: avoid accept() giving EMFILE
When enough connections were opened with the target server so that all
available file descriptors were consumed, the server would enter
an infinite loop which would cause an error message to be printed to the log
file at a rapid rate, causing 100% CPU utilisation.

Avoid this case by introducing a configurable limit of open file
descriptors that the library is allowed to make use of before denying
new client connections.