Use binary encoding as default instead of base64
The WebSocket specification does not include base64 encoding. It was
mentioned in a very early draft of the specification but never made it
to any final version.

Not many WebSocket clients out there support base64 data, and the ones
that do are very outdated.

Then there's the topic of subprotocols - Prior to this commit,
libvncserver required a client to specify 'binary' as a subprotocol in
order to not get base 64 data. This is also a non-standard value.

With this commit the code is still compatible with old clients that
specifically ask for 'binary' or 'base64' using subprotocols.