Use platform-specific type for socket FDs
The Windows Sockets 2 API uses the type SOCKET for all functions and is
defined as an unsigned integer with an architecture-specific width.
Storing the result of WS2 functions in an integer field is therefore
insufficient and may lead to a wrong behaviour. Additionally invalid
sockets are not represented by negative numbers but the constant
INVALID_SOCKETS.

This commit wraps the platform-specific socket type and a constant for
an invalid socket.