rfbproto: fix broken endianess handling on Win32
It's definitely wrong to build with LIBVNCSERVER_WORDS_BIGENDIAN
defined on Win32. Instead use the auto-detected endianess macro
(which will indicate little endian on i686/x86_64). Removing the
hardcoded define makes all tests pass (especially encodingtest)
when built with the MinGW64 toolchain.

In combination with PR #257 this fixes issues #165 and #249.