libvncclient: rfbproto: do not fallback to IPv4
The IPv4 code which is used as fallback contains calls to deprecated
network functions such as gethostbyname() which are not thread-safe
and thus can lead to crashes when called in multiple threads at the
same time. The IPv6 code properly uses getaddrinfo() which does not
suffer from these problems. When IPv6 is enabled the modern code
paths should be used exclusively. It's unlikely anyway that the
legacy network code succeeds after IPv6 (which is capable of
establishing IPv4 connections as well) failed before.

Closes #262.