LibVNCClient: add connect timeout support
Depending on the platform and the TCP/IP stack configuration connect()
can block for a very long time if the destination host is unreachable.
The current behaviour blocks VNC client loops for an uncertain amount of
time. However some applications may want to implement responsive VNC
client threads requiring rfbInitClient() to time out more quickly. This
can be achieved by calling connect() on a non-blocking socket and
select() with a timeout afterwards. The new WaitForConnected() function
deals with the platform-specific behaviour of this mechanism.