rfb/rfb.h: increase UPDATE_BUF_SIZE by 2768 bytes
Simply raise the default to 32768 which is not an awful lot of an
increase and should not hurt the embedded guys.

If someone needs more, we can still go down the malloc() route and
make the buffer size configurable, but leave this to the users to
request and KISS for now.

Closes #521