libvncclient: replace gethostbyname() with getaddrinfo()
man gethostbyname:
"The gethostbyname*(), gethostbyaddr*(), herror(), and hstrerror() functions are obsolete.  Applications should use getaddrinfo(3), getnameinfo(3), and gai_strerror(3) instead."

Furthermore, the man page states that `gethostbyname()` is MT-Unsafe, probably due to the fact that it retuns a pointer to static data. Thus, by using `getaddrinfo()`, we improve the multithreading soundness of LibVNCClient.