libvncclient: support all tight 32bpp formats
Prior to this change, the only 32bpp pixel formats that the tight
decoder supported were the ABGR, BGRA, ARGB, and RGBA 32bpp pixel
formats. This is because these are the only RGB formats that TurboJPEG
can directly output.  To support all possible 32bpp pixel formats with
all possible maximums, shifts, etc., we need to perform additional
conversion after decompressing the JPEG.  Thankfully the 16bpp code path
is already doing this, so we can reuse it if the client requests a 32bpp
pixel format that TurboJPEG cannot directly output.