test/encodingstest: support more bpps and sizes
Add the options -c8, -c16, and -c32 to test with client bpps of 8, 16,
and 32, respectively.  Also add options -s8, s16, and -s32 to test with
server bpps of 8, 16, and 32, respectively.  By default, the client and
server are tested with a bpp of 32, preserving the original behavior
when no bpp options are passed.

Also add the -large option.  Normally, a 400x300 image is tested, but
when -large is passed a 1024x768 image is tested.